:recycle: Extract issue details action to dedicated component
